requirements.txt
flask
supabase
requests
groq
dotenv
jsonify
markdown-it-py
mdit-py-plugins
linkify-it-py